requirements.txt
certifi>=2020.12.5
cycler>=0.10.0
kiwisolver>=1.3.1
matplotlib>=3.3.4
numpy>=1.19.5
Pillow>=8.1.0
pyparsing>=2.4.7
python-dateutil>=2.8.1
PyYAML>=5.4.1
scipy>=1.5.4
six>=1.15.0
